doorbell-pi/requirements.txt
gpiozero
numpy
tqdm
python-telegram-bot
configargparse